Initial commit of Async/SyncMcpToolGroupProviders
This providers take on construction an Object and a Class<?> that
represents a tool group...i.e. a grouping of MCP tool methods, assumed
to be annotated with McpTool annotations.   From the given Class<?> the
provider generates a List of async/sync tool descriptions + method
callbacks on the serviceObject (assumed/required to implement the given
tool group class.