Introduce tool groups (explicit classes/interfaces) to generalize tool providers
This commit is for the SyncMcpToolProvider class only (not async or
stateless).  Once the general idea has been agreed upon, I will
similarly modify the other provider classes.

The basic idea is to generalize the *ToolProvider constructor to allow
the specification generation to use java interfaces and classes to use
as McpTool annotation inspection.  For example:

public SyncMcpToolProvider(List toolObjects, Class<?>... toolGroups) {
...
}

This adds a Class<?>[] that will be used rather than the
toolObject.getClass() call to reflect on the methods for finding
@McpTool annotations.

This allows a usage like this:

provider = new SyncMcpToolProvider(myObject, MyInterface.class)

A usage like this would be enabling for environments like OSGi services,
where the instance only is required to implement MyInterface.  That's my
specific use case...OSGi services/remote services.

Please see this example:

This implies that MyInterface is declared like this

interface MyInterface {
    @McpTool(name="tool1", description="An LLM friendly description of
this method/tool"
    void myMethod();
}

Using the Class<?>... in the constructor has the nice property that the
previous constructor (i.e. List toolObject) can still be used exactly
the same:

provider = new SyncMcpToolProvider(List.of(toolObject));

Meaning that it allows the same/bean usage as originally implemented
(which uses the toolObject's clazz and clazz.getDeclaredMethods() to
inspect methods).

But by having the Class<?>... it allows other usage patterns...like OSGi
services without changing anything about the existing provider behavior.

This can be verified via the tests for this provider...which work
exactly as before.

I've added some tests for the toolGroups usage to the Test class as part
of this commit.  If others are desired, I can and will add them.

Signed-off-by: Scott Lewis <scottslewis@gmail.com>